pyproject.toml
[project]
name = "openserver"
version = "0.0.1"
description = "Open Source TheProtocols Server"
readme = "README.md"
license = {text = "GPL-3.0"}
maintainers = [{name = "Çağan Mert İŞLEK", email = "islekcaganmert@gmail.com"}]
classifiers = [
    "License :: OSI Approved :: GPL-3.0 License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
]
requires-python = ">=3.10"
dependencies = ["requests", "TheProtocols", "flask[async]", "pyyaml", "cryptography", "flask-sqlalchemy", "beautifulsoup4", "asyncio"]

[project.urls]
Donate = "https://www.buymeacoffee.com/islekcaganmert"
Documentation = "https://github.com/islekcaganmert/openserver/wiki"
Changes = "https://github.com/islekcaganmert/openserver/commits/main"
"Source Code" = "https://github.com/islekcaganmert/openserver"
"Issue Tracker" = "https://github.com/islekcaganmert/openserver/issues"
Chat = "https://lemmy.today/c/openserver"

[build-system]
requires = ["setuptools","wheel"]
build-backend = "setuptools.build_meta"